refactor(fav-route): changed name of fav route and added empty state to view
