+++ README.md
@@ -5,7 +5,7 @@
 
 A mobile application built with Expo, displaying random quotes and a paginated list of quotes. It includes:
 
-- **Expo Router** – bottom-tabs navigation with Home / Quotes / Favorites.
+- **Expo Router** – bottom-tabs navigation with Home / Quotes / Favourites.
 - **React Query** (`@tanstack/react-query`) – infinite scroll, pull-to-refresh, state handling
 - **Axios** – typed client (API layer in `/services`).
 - **Global Theme** (light/dark) + `AppThemeSwitcher` with animated icon.
@@ -24,9 +24,9 @@ A mobile application built with Expo, displaying random quotes and a paginated l
 
 | Screen             | Functionality                                                                                                   |
 | ------------------ | --------------------------------------------------------------------------------------------------------------- |
-| **Home**           | Random quote (fetch), _New quote_ / _Add to favorites_ buttons                                                  |
+| **Home**           | Random quote (fetch), _New quote_ / _Add to favourites_ buttons                                                 |
 | **Quotes**         | Infinite-scroll list, search bar (debounced with `router.setParams`), pull-to-refresh, loading indicator/footer |
-| **Favorites**      | Placeholder for saved quotes                                                                                    |
+| **Favourites**     | Placeholder for saved quotes                                                                                    |
 | **Theme switcher** | Animated sun/moon icon rotation, `ThemeProvider` context                                                        |
 
 ---
